+++ core/services/ocr2/plugins/ocr2keeper/evm21/logprovider/basetypes_copy.go
@@ -1,4 +1,4 @@
-package evm
+package logprovider
 
 import (
 	"crypto/sha256"
