+++ core/services/ocr2/plugins/ocr2keeper/evm21/logprovider/buffer.go
@@ -1,4 +1,4 @@
-package evm
+package logprovider
 
 import (
 	"math/big"
